+++ agent-app/src/apis/connectAPI.js
@@ -1,13 +1,13 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT-0
 
-import { API } from '@aws-amplify/api';
+import { RestAPI } from '@aws-amplify/api-rest';
 import { Auth } from '@aws-amplify/auth';
 
 export const ccpLogin = async (connectLoginByEmail) => {
     //get current Cognito ID Token
     let cid = (await Auth.currentSession()).getIdToken().getJwtToken();
-    const response = await API.post('videoCallEscalationConnectAPI', '/ccplogin', {
+    const response = await RestAPI.post('videoCallEscalationConnectAPI', '/ccplogin', {
         headers: {
             cognitoIdToken: cid
         },