+++ agent-app/webpack.config.dev.js
@@ -6,7 +6,6 @@ module.exports = merge(commonConfig, {
   devtool: 'source-map',
   output : { publicPath : '/'},
   devServer: {
-            open: true,
-            openPage: 'demo-website.html'
+            open: true
         }
 })