+++ demo-website/webpack.config.common.js
@@ -1,35 +1,23 @@
 const path = require('path');
-
-const HtmlWebpackInlineSourcePlugin = require('html-webpack-inline-source-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const RemovePlugin = require('remove-files-webpack-plugin');
+const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 
 const app = process.env.npm_config_app || 'demo_website'
 
 module.exports = {
         plugins: [
+            new CleanWebpackPlugin({verbose: true}),
             new HtmlWebpackPlugin({
-                inlineSource: '.(js|css)$',
                 template: __dirname + `/src/index.html`,
                 filename: __dirname + `/build/demo-website.html`,
+                inlineSource: '.(css)$',
                 inject: 'head'
-            }),
-            new HtmlWebpackInlineSourcePlugin(),
-            new RemovePlugin({
-                after: {
-                    include: ['./build/demo-bundle.js']
-                }
             })
         ],
-        entry: ['babel-polyfill', path.resolve(__dirname, 'src') + '/meeting.js'],
-        resolve: {
-            extensions: ['.webpack.js', '.web.js', '.js'],
-        },
+        entry: [path.resolve(__dirname, 'src') + '/meeting.js'],
         output: {
             path: path.resolve(__dirname, 'build'),
-            filename: 'demo-bundle.js',
-            libraryTarget: 'var',
-            library: `app_${app}`,
+            filename: 'demo-bundle.js'
         },
         module: {
             rules: [
@@ -79,7 +67,7 @@ module.exports = {
                         loader: 'babel-loader',
                         options: {
                             presets: ['@babel/preset-env'],
-                            plugins: ['@babel/plugin-proposal-class-properties']
+                            plugins: ['@babel/plugin-proposal-class-properties', "@babel/transform-runtime"]
                         }
                     }
                 }