+++ CHANGELOG.md
@@ -4,7 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2021-02-19
+### Added
+ - [agent-app] webpack and babel packages
+ - [agent-app] aws-amplify/api-rest package
+ - [agent-app] webpack config files
+ - [demo-website] babel/runtime and babel/plugin-transform-runtime packages
+### Changed
+ - [agent-app] aws-amplify/auth, aws-amplify/ui-react packages
+ - [agent-app] start and build npm scripts
+ - [agent-app] chimeAPI.js and connectAPI.js now use aws-amplify/api-rest
+ - [demo-website] popper.js, webpack, *-loader packages
+### Removed
+ - [agent-app] react-scripts, aws-amplify/core, aws-amplify/api, jest packages
+ - [agent-app] robots.txt and favicon.ico
+ - [demo-website] babel-polyfill, remove-files-webpack-plugin, html-webpack-inline-source-plugin packages
+ - [demo-website] start:dev npm script
 
+
+
+### Fixed
 ## [1.0.0] - 2021-02-16
 ### Added
 - Initial import